cli: show service status in tunnel log
Fixes #183714

+++ cli/src/util/zipper.rs
@@ -88,8 +88,7 @@ where
 			use std::io::Read;
 			use std::os::unix::ffi::OsStringExt;
 
-			if matches!(file.unix_mode(), Some(mode) if mode & (S_IFLNK as u32) == (S_IFLNK as u32))
-			{
+			if matches!(file.unix_mode(), Some(mode) if mode & S_IFLNK == S_IFLNK) {
 				let mut link_to = Vec::new();
 				file.read_to_end(&mut link_to).map_err(|e| {
 					wrap(
